rune/skeleton: fix liberpal-skeleton-v3 crash
EPM feature is only supported by newer than PAL v3. However,
current implementation deems a PAL without EPM configured as
PAL v1. This causes a PAL v3 without EPM configured calls
PAL v1 init().
